Репозиторий Серверов в памяти
